DB-management/requirements.txt
pymysql
requests
tqdm
google-cloud-translate